changed unsigned to size_t

+++ src/main.cpp
@@ -236,7 +236,7 @@ vector<string> tokenize(string x)
 vector<string> separateByDelim(vector<string> origVector, string delimiter)
 {
     vector<string> tokenVectorFinal;
-    unsigned pos = 0;
+    size_t pos = 0;
     string oldString;
     string newString;
     for(unsigned l = 0; l < origVector.size(); l++)
